dependencies/win/requirements.txt
PyYAML==5.4.1
patool==1.12
schema==0.7.5